Parameters enabled!!!
Now is posible send via parameters the url that will be translated and where de scraper will search the translated strings.

An example of that would be:

A basic call of the scraper
scrapy crawl getStrings

add to that the scrapy parameter "-a" with the word "toTranslate=" and next to it the url.

A full example
scrapy crawl getStrings -a toTranslate="https://xxx.xxx" -a translated="https://xxx.xxx" -o woocomerse.csv -t csv